CMake update and associated warning fixes.
- Deprecation warnings are disabled due to use of deprecated CUDA APIs
  such as cudaBindTexture. Filed NVlabs/cub#191 to track this.
- Bump CMake minimum version to 3.15.
- Remove workarounds for old cmake/compilers.
- Switch to CUB_ENABLE_DIALECT_CPPXX instead of CMAKE_CXX_STANDARD.
  - Like Thrust multiconfig, all can be enabled at once.
  - Uses Thrust's settings when Thrust multiconfig is enabled.
- Add CUB_ENABLE_* options to disable portions of the build.
  - HEADER_TESTING, TESTING, EXAMPLES
- Add sm_37, sm_80, remove sm_30, sm_32.
  - When building CUB in Thrust, just use thrust's flags.
- Sanitize variable names in AppendOptionIfAvailable.
  - GCC was failing to detect `-Werror`. Long story short, this fixes it.
- Add cub.compiler_interface target to store warning suppressions, etc.
- Find and use Thrust. We should move the Thrust code into Thrust.
- Check for THOROUGH, QUICK and QUICKER variants of tests, create new
  targets for them when found.
- Remove `examples/block/reduce_by_key.cu`, since it doesn't build and is
  incomplete.
- Disable NPP code paths in test_device_histogram as they don't compile.
  Seems like a Dispatch<int, int, NPP> specialization got removed at some
  point.
- Misc warning cleanups.
- Add all project files to custom target for IDE detection.

+++ test/mersenne.h
@@ -59,7 +59,7 @@ static int mti = N + 1;     /* mti==N+1 means mt[N] is not initialized */
 void init_genrand(unsigned int s)
 {
     mt[0] = s & 0xffffffff;
-    for (mti = 1; mti < N; mti++)
+    for (mti = 1; mti < static_cast<int>(N); mti++)
     {
         mt[mti] = (1812433253 * (mt[mti - 1] ^ (mt[mti - 1] >> 30)) + mti);
 
@@ -83,7 +83,9 @@ void init_by_array(unsigned int init_key[], int key_length)
     init_genrand(19650218);
     i = 1;
     j = 0;
-    k = (N > key_length ? N : key_length);
+    k = (static_cast<int>(N) > key_length
+	 ? static_cast<int>(N)
+	 : key_length);
     for (; k; k--)
     {
         mt[i] = (mt[i] ^ ((mt[i - 1] ^ (mt[i - 1] >> 30)) * 1664525))
@@ -91,7 +93,7 @@ void init_by_array(unsigned int init_key[], int key_length)
         mt[i] &= 0xffffffff;    /* for WORDSIZE > 32 machines */
         i++;
         j++;
-        if (i >= N)
+        if (i >= static_cast<int>(N))
         {
             mt[0] = mt[N - 1];
             i = 1;
@@ -103,7 +105,7 @@ void init_by_array(unsigned int init_key[], int key_length)
         mt[i] = (mt[i] ^ ((mt[i - 1] ^ (mt[i - 1] >> 30)) * 1566083941)) - i; /* non linear */
         mt[i] &= 0xffffffff; /* for WORDSIZE > 32 machines */
         i++;
-        if (i >= N)
+        if (i >= static_cast<int>(N))
         {
             mt[0] = mt[N - 1];
             i = 1;
@@ -121,19 +123,19 @@ unsigned int genrand_int32(void)
 
     /* mag01[x] = x * MATRIX_A  for x=0,1 */
 
-    if (mti >= N)
+    if (mti >= static_cast<int>(N))
     { /* generate N words at one time */
         int kk;
 
         if (mti == N + 1) /* if init_genrand() has not been called, */
         init_genrand(5489); /* a defat initial seed is used */
 
-        for (kk = 0; kk < N - M; kk++)
+        for (kk = 0; kk < static_cast<int>(N - M); kk++)
         {
             y = (mt[kk] & UPPER_MASK) | (mt[kk + 1] & LOWER_MASK);
             mt[kk] = mt[kk + M] ^ (y >> 1) ^ mag01[y & 0x1];
         }
-        for (; kk < N - 1; kk++)
+        for (; kk < static_cast<int>(N - 1); kk++)
         {
             y = (mt[kk] & UPPER_MASK) | (mt[kk + 1] & LOWER_MASK);
             mt[kk] = mt[kk + (M - N)] ^ (y >> 1) ^ mag01[y & 0x1];
